proper int constant types

+++ src/main.rs
@@ -1,5 +1,4 @@
 use std::ffi::CString;
-use std::os::raw::c_int;
 use std::ptr::null_mut;
 
 use scopeguard::defer;
@@ -11,9 +10,9 @@ fn main() {
     c::glfwInit();
     defer!(c::glfwTerminate());
 
-    c::glfwWindowHint(c::GLFW_CONTEXT_VERSION_MAJOR as c_int, 4);
-    c::glfwWindowHint(c::GLFW_CONTEXT_VERSION_MINOR as c_int, 6);
-    c::glfwWindowHint(c::GLFW_OPENGL_PROFILE as c_int, c::GLFW_OPENGL_CORE_PROFILE as c_int);
+    c::glfwWindowHint(c::GLFW_CONTEXT_VERSION_MAJOR, 4);
+    c::glfwWindowHint(c::GLFW_CONTEXT_VERSION_MINOR, 6);
+    c::glfwWindowHint(c::GLFW_OPENGL_PROFILE, c::GLFW_OPENGL_CORE_PROFILE);
 
     let window_title = CString::new("hello world").unwrap();
     let window = c::glfwCreateWindow(640, 360, window_title.as_ptr(), null_mut(), null_mut());
@@ -23,7 +22,7 @@ fn main() {
     c::glfwSwapInterval(1);
     c::gladLoadGL(Some(c::glfwGetProcAddress));
 
-    while c::glfwWindowShouldClose(window) == c::GLFW_FALSE as c_int {
+    while c::glfwWindowShouldClose(window) == c::GLFW_FALSE {
       // https://github.com/rust-lang/rust-bindgen/issues/753
       c::glad_glClearColor.unwrap()(0.7, 0.9, 0.1, 1.0);
       c::glad_glClear.unwrap()(c::GL_COLOR_BUFFER_BIT);
